+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.3.0
+
+### Fixes
+- Fix the error message when an `ApiException` is thrown and use the `ProblemDetails.Title` if possible.
+- Add more properties to the `ProblemDetails` type to more accurately represent the response.
+- **[BREAKING]** Types `ApiException` and `ProblemDetails` have been moved to namespace `Laserfiche.Api.Client`.
+- **[BREAKING]** Property `ProblemDetails.AdditionalProperties` has been removed. Use property `ProblemDetails.Extensions` instead.
+- **[BREAKING]** Types of `ApiException<T>` has been removed. Use `ApiException` instead. The `ApiException` has a `ProblemDetails` property which may contain additional information.
+
 ## 1.2.3
 
 ### Fixes